backend/requirements.txt
flask
flask-cors
groq
python-dotenv
pydantic
uvicorn
asgiref
